* libltdl/ltdl.c (load_deplibs, try_dlopen): Revert unnecessary casts to int from my patch from 2004-10-01.

+++ libltdl/ltdl.c
@@ -737,10 +737,10 @@ load_deplibs (lt_dlhandle handle, char *deplibs)
   p = deplibs;
   while (*p)
     {
-      if (!isspace ((int)(unsigned char) *p))
+      if (!isspace ((unsigned char) *p))
 	{
 	  char *end = p+1;
-	  while (*end && !isspace((int)(unsigned char) *end))
+	  while (*end && !isspace((unsigned char) *end))
 	    {
 	      ++end;
 	    }
@@ -786,14 +786,14 @@ load_deplibs (lt_dlhandle handle, char *deplibs)
   p = deplibs;
   while (*p)
     {
-      if (isspace ((int)(unsigned char) *p))
+      if (isspace ((unsigned char) *p))
 	{
 	  ++p;
 	}
       else
 	{
 	  char *end = p+1;
-	  while (*end && !isspace ((int)(unsigned char) *end))
+	  while (*end && !isspace ((unsigned char) *end))
 	    {
 	      ++end;
 	    }
@@ -1004,7 +1004,7 @@ try_dlopen (lt_dlhandle *phandle, const char *filename)
     int i;
     for (i = 0; i < ext - base_name; ++i)
       {
-	if (isalnum ((int)(unsigned char)(base_name[i])))
+	if (isalnum ((unsigned char)(base_name[i])))
 	  {
 	    name[i] = base_name[i];
 	  }
